Change parseInt() to parseFloat()

+++ script.js
@@ -52,16 +52,16 @@ const calculate = () => {
     let result = '';
     switch(calculationOperator) {
         case '+':
-            result = parseInt(prevNumber) + parseInt(currentNumber);
+            result = parseFloat(prevNumber) + parseFloat(currentNumber);
             break;
         case '-':
-            result = parseInt(prevNumber) - parseInt(currentNumber);
+            result = parseFloat(prevNumber) - parseFloat(currentNumber);
             break;
         case '*':
-            result = parseInt(prevNumber) * parseInt(currentNumber);
+            result = parseFloat(prevNumber) * parseFloat(currentNumber);
             break;
         case '/':
-            result = parseInt(prevNumber) / parseInt(currentNumber);
+            result = parseFloat(prevNumber) / parseFloat(currentNumber);
             break;
         default:
             return;
